+++ NEWS.md
@@ -2,6 +2,10 @@
 
 ## Changes compared to last release
 
+- Bugfix in OpenAI metadata extraction
+- New `ollama_delete_model()` function
+- `list_models()` is now a verb supporting both `openai()` and `ollama()`. Other providers will be added
+- Ellmer schemata supported in `.json_schema` arguements
 - New `send_ollama_batch()` function to make use of the fast parallel request features of Ollama. 
 - New batch functions for Azure Openai (thanks [Jia Zhang](https://github.com/JiaZhang42))
 - New parameters for `openai()` reasoning models supported